Move Fetchable and Bidirectional to ObjectService
