+++ java-client/session-dagger/src/test/java/io/deephaven/client/ObjectServiceTest.java
@@ -8,8 +8,8 @@ import io.deephaven.client.impl.DataAndTypedTickets;
 import io.deephaven.client.impl.HasTypedTicket;
 import io.deephaven.client.impl.ObjectService.MessageStream;
 import io.deephaven.client.impl.ServerObject;
-import io.deephaven.client.impl.ServerObject.Bidirectional;
-import io.deephaven.client.impl.ServerObject.Fetchable;
+import io.deephaven.client.impl.ObjectService.Bidirectional;
+import io.deephaven.client.impl.ObjectService.Fetchable;
 import io.deephaven.client.impl.TableHandle;
 import io.deephaven.client.impl.TableHandle.TableHandleException;
 import io.deephaven.client.impl.TableObject;
